+++ modules/iam-service-account/outputs.tf
@@ -16,12 +16,12 @@
 
 output "service_account" {
   description = "Service account resource."
-  value       = google_service_account.service_account
+  value       = local.service_account
 }
 
 output "email" {
   description = "Service account email."
-  value       = google_service_account.service_account.email
+  value       = local.service_account.email
 }
 
 output "iam_email" {